Delete unused models and refactor the namespaces structure
This commit deletes two models: BackdropSize and LangId, which were no longer being used in the codebase. Also, several models are moved to different directories as part of a comprehensive namespace structure refactor to improve code organization. This includes forming separate namespaces for Data, Constants, Enums, Api, and Configs. This will make future maintenance easier and the structure more understandable. Several using directives have also been updated to match the new structure.
