--- FoliCon/Models/BackdropSize.cs
@@ -1,9 +1,0 @@
-﻿namespace FoliCon.Models;
-
-public static class BackdropSize
-{
-    public static string W300 { get; } = "w300";
-    public static string W780 { get; } = "w780";
-    public static string W1280 { get; } = "w1280";
-    public static string Original { get; } = "original";
-}
