+++ FoliCon/Models/Data/ListViewData.cs
@@ -1,4 +1,4 @@
-﻿namespace FoliCon.Models;
+﻿namespace FoliCon.Models.Data;
 
     public class ListViewData : BindableBase
     {
