--- FoliCon/Models/LangId.cs
@@ -1,11 +1,0 @@
-﻿namespace FoliCon.Models;
-
-//https://www.autoitscript.com/autoit3/docs/appendix/OSLangCodes.htm
-public enum LangId
-{
-    English = 1033,
-    Arabic = 1025,
-    Spanish = 1034,
-    Hindi = 1081,
-    Russian = 1049
-}
