+++ FoliCon/Modules/IGDB/IgdbDataTransformer.cs
@@ -1,4 +1,5 @@
-﻿using FoliCon.Modules.utils;
+﻿using FoliCon.Models.Data;
+using FoliCon.Modules.utils;
 using NLog;
 using Logger = NLog.Logger;
