+++ FoliCon/Modules/TMDB/TMDB.cs
@@ -1,4 +1,5 @@
-﻿using NLog;
+﻿using FoliCon.Models.Data;
+using NLog;
 using Collection = TMDbLib.Objects.Collections.Collection;
 using Logger = NLog.Logger;
