+++ FoliCon/Modules/TMDB/TMDBService.cs
@@ -1,4 +1,6 @@
-﻿using NLog;
+﻿using FoliCon.Models.Constants;
+using FoliCon.Models.Data;
+using NLog;
 
 namespace FoliCon.Modules.TMDB;
