+++ FoliCon/Modules/TMDB/TmdbDataTransformer.cs
@@ -1,4 +1,6 @@
-﻿using FoliCon.Modules.utils;
+﻿using FoliCon.Models.Constants;
+using FoliCon.Models.Data;
+using FoliCon.Modules.utils;
 using NLog;
 using Collection = TMDbLib.Objects.Collections.Collection;
