+++ FoliCon/Modules/utils/FileUtils.cs
@@ -1,4 +1,6 @@
-﻿using FoliCon.Modules.Configuration;
+﻿using FoliCon.Models.Constants;
+using FoliCon.Models.Data;
+using FoliCon.Modules.Configuration;
 using FoliCon.Modules.IGDB;
 using FoliCon.Modules.TMDB;
 using FoliCon.Modules.UI;
