+++ FoliCon/ViewModels/SearchResultViewModel.cs
@@ -1,4 +1,6 @@
-﻿using FoliCon.Modules.Extension;
+﻿using FoliCon.Models.Constants;
+using FoliCon.Models.Data;
+using FoliCon.Modules.Extension;
 using FoliCon.Modules.IGDB;
 using FoliCon.Modules.TMDB;
 using FoliCon.Modules.UI;
